fixing catalogueOfUser function
