+++ components/userCatalogue/index.tsx
@@ -1,6 +1,6 @@
 import React, { useEffect, useState } from "react";
 import { MostrarProductos, Root } from "./styled";
-import { getUserCatalogue } from "lib/api";
+import { getUserCatalogue } from "lib/hooks";
 import { Pagination } from "ui/pagination";
 import { useRouter } from "next/router";
 import { PageButton } from "ui/buttons";
