+++ lib/api.ts
@@ -70,14 +70,3 @@ export function removeItem(item: string) {
 }
 export { fetchApi, sendCode, getToken };
 
-export async function getUserCatalogue(userId: string) {
-
-	if (userId) {
-		const catalogue = await fetchApi(`/catalogue?q=${userId}&offset=0&limit=10`, {
-			method: "POST",
-			headers: { "Content-Type": "application/json" },
-		});
-		console.log(catalogue);
-		return catalogue;
-	}
-}
