+++ lib/hooks.ts
@@ -81,3 +81,13 @@ export async function getOrder(externalReference: any) {
     : false;
 }
 
+export async function getUserCatalogue(userId: string) {
+
+  if (userId) {
+    const catalogue = await fetchApi(`/catalogue?q=${userId}&offset=0&limit=10`, {
+      method: "POST",
+      headers: { "Content-Type": "application/json" },
+    });
+    return catalogue;
+  }
+}
