+++ pages/catalogue/[userId].tsx
@@ -2,7 +2,7 @@ import { ShowCatalogue } from "components/userCatalogue";
 import { Header } from "components/header";
 import { Footer } from "components/footer";
 import type { NextPage } from "next";
-import { Root } from "./styled";
+import Root from "./styled";
 
 const UserCatalog: NextPage = () => {
 	return (
