+++ pages/catalogue/styled.ts
@@ -1,9 +1,11 @@
 import styled from "styled-components";
 
-export const Root = styled.div`
+const Root = styled.div`
 	justify-content: center;
 	flex-direction: column;
 	align-items: center;
 	min-height: 800px;
 	display: flex;
 `;
+
+export default Root;
